refactor(reborn-cli): address review feedback — split repl runtime helper (#3785)
