+++ crates/ironclaw_reborn_cli/src/main.rs
@@ -1,6 +1,7 @@
 mod cli;
 mod commands;
 mod context;
+mod runtime;
 
 fn main() -> anyhow::Result<()> {
     cli::run()
